Fixed signed/unsigned bug from prev commit

+++ src/media_tools/dash_client.c
@@ -449,7 +449,7 @@ static void gf_dash_group_timeline_setup(GF_MPD *mpd, GF_DASH_Group *group, u64 
 	}
 
 	availabilityStartTime = 0;
-	if (mpd->availabilityStartTime + group->dash->utc_shift > - group->dash->utc_drift_estimate) {
+	if ((s64) mpd->availabilityStartTime + group->dash->utc_shift > (s64) - group->dash->utc_drift_estimate) {
 		availabilityStartTime = mpd->availabilityStartTime + group->dash->utc_shift + group->dash->utc_drift_estimate;
 	}
